Avoid the page getting in a bad state.
Occasionally after it running for a long time the page stops refreshing
and the boxes are filled with ‘?’. I think this may be due to a JS
error, possibly from handling of failed AJAX requests.

This makes the page refresh every hour. This feels like the easiest
thing to do to since replicating the error would take up a lot of time.