Fix bug that occurs when latest commit is authored using `git-pair`
If the latest commit is authored by two people using git-pair [1], the
`repo_state` object returned by the Github API does not contain information
about the author of the commit. This was resulting in JS errors, preventing the
SHA from being displayed.

This commit fixes this bug by checking if the required commit author information
is available first before trying to access it.

[1] https://github.com/chrisk/git-pair